Add option for and keep track of pathsfifo
The fd and path to the named pipe are stored in the Vis struct and
initalized when the struct is created.

The pathsfifo option, to the user, specifies the path to the named pipe.
When this option is set, the named pipe is either opened or created. If
that is sucessful, then any previous values are freed and a message is
displayed.