Add closing of old windows and opening of new ones
Before reading paths, all closable windows (ones that :qa would close;
unmodified and not internal) are closed. If it cannot be closed, its
path is noted so that the maximum number of open files is not exceeded.

After reading, all new paths are opened (as if each was called with
:split/:vsplit).